Load cart from local storage
